# src/router/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": " IANode ACCESS",
    "version": "1.0.0",
    "description": "Integration Architecture Node ACCESS API",
    "contact": {
      "name": "IANode",
      "url": "https://ndtp.co.uk/",
      "email": "NDTP@businessandtrade.gov.uk"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8091"
    }
  ],
  "paths": {
...
    "/groups": {
      "get": {
        "summary": "Get all groups",
        "tags": ["Groups"],
        "description": "Returns all groups",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully retrieved array of groups",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Groups not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupsNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new group",
        "tags": ["Groups"],
        "description": "Creates a new group",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroup"
              }
            }
          }
        },
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "Created group ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGroupSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request / Fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupBadRequest"
                }
              }
            }
          },
          "409": {
            "description": "Group already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupAlreadyExists"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}": {
      "get": {
        "summary": "Get specific group by group_id",
        "tags": ["Groups"],
        "description": "Get specific group by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "description": "unique identifier for groups"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully retrieved group",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Group"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "example": [
                            {
                              "_id": "6540c6f026bc7a0bd665849f",
                              "id": "6540c6f026bc7a0bd665849f",
                              "name": "Example Name",
                              "active": true
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupsNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete specific group by group_id",
        "tags": ["Groups"],
        "description": "Delete group from access - soft delete which removes the group from use by updating it to be inactive",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "description": "unique identifier for groups"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully soft deleted group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteGroupSuccess"
                }
              }
            }
          },
          "404": {
            "description": "Group not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupsNotFound"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Group": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "6540c84426bc7a0bd66584aa"
          },
          "group_id": {
            "type": "string",
            "example": "urn:ndtp:groups:example"
          },
          "label": {
            "type": "string",
            "example": "Example group name"
          },
          "description": {
            "type": "string",
            "example": "A group defined as an example"
          },
          "active": {
            "type": "boolean",
            "default": false
          },
          "userCount": {
            "type": "number",
            "required": false,
            "example": 1
          }
        }
      },
      "CreateGroup": {
        "type": "object",
        "required": ["label", "description", "active"],
        "properties": {
          "label": {
            "type": "string",
            "example": "example"
          },
          "description": {
            "type": "string",
            "example": "A group defined as an example"
          },
          "active": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "CreateGroupSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "created": {
                "type": "boolean",
                "example": true
              },
              "uuid": {
                "type": "string",
                "example": "urn:ndtp:groups:example"
              }
            }
          }
        }
      },
      "DeleteGroupSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "deleted": {
                "type": "boolean",
                "example": true
              }
            }
          }
        }
      },
      "GroupBadRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Invalid request / Fields missing"
          }
        }
      },
      "GroupsNotFound": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Group(s) not found"
          }
        }
      },
      "GroupAlreadyExists": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Group already exists"
          }
        }
      },
...
      "ServerError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 500
          },
          "message": {
            "type": "string",
            "example": "Server error"
          }
        }
      },
...
    }
  },
...
}